[executor] Rewrite executor.
This commit rewrites executor significantly.

Now, field execution in an Object becomes a task. The tasks are sent to
an executor to schedule for execution. An executor can optionally be
backed by a `concurrent.Executor` (see `executor.ExecuteParams`)
introduced in an earlier commit to enable parallel execution.

`executor.PreparedOperation.Execute` becomes a non-blocking call
which returns an `<-chan ExecutionResult`. But when
`concurrent.Executor` is not supplied, it is still a blocking call
because the execution is performed by the calling goroutine.

`execute_test.go` is also refactored to test with and without
`concurrent.Executor`.
